+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/ratify-project/ratify-cli/v2
 
-go 1.23.0
+go 1.23.1
 
 require github.com/spf13/cobra v1.8.1
